% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\docType{class}
\name{class-isoTAX}
\alias{class-isoTAX}
\title{isoTAX Class Object}
\value{
Returns an class-isoTAX object.
}
\description{
S4 wrapper for \code{\link{isoTAX}} function. Access data via S4 slot functions.
}
\section{Slots}{

\describe{
\item{\code{input}}{Character string containing input directory information.}

\item{\code{warning}}{Character string containing list filenames of sequences that had poor alignment during taxonomic classification step.}

\item{\code{date}}{Character string containing run date from each of the input Sanger sequence .ab1 files ("YYYY_MM_DD" format).}

\item{\code{filename}}{Character string containing input filenames.}

\item{\code{phred_spark_raw}}{List containing per nucleotide Phred score values for each sequence}

\item{\code{phred_raw}}{Numeric string containing mean Phred scores before trimming.}

\item{\code{phred_trim}}{Numeric string containing mean Phred scores after trimming.}

\item{\code{Ns_raw}}{Numeric string containing count of N's before trimming.}

\item{\code{Ns_trim}}{Numeric string containing count of N's after trimming.}

\item{\code{length_raw}}{Numeric string containing sequence length before trimming.}

\item{\code{length_trim}}{Numeric string containing sequence length after trimming.}

\item{\code{seqs_raw}}{Character string containing sequences before trimming.}

\item{\code{seqs_trim}}{Character string containing sequence after trimming.}

\item{\code{closest_match}}{Character string containing species + type strain no. of closest match from reference database.}

\item{\code{NCBI_acc}}{Character string containing NCBI accession number associated with closest match from reference database.}

\item{\code{ID}}{Numeric string containing containing pairwise similarity value for query vs database reference sequence.
Calculation of ID is determined by isoTAX 'iddef' parameter (0-4, Default=2). See VSEARCH documentation for more details.
\itemize{
\item (0) CD-HIT definition: (matching columns) / (shortest sequence length).
\item (1) Edit distance: (matching columns) / (alignment length).
\item (2) Edit distance excluding terminal gaps (default definition).
\item (3) Marine Biological Lab definition counting each gap opening (internal or terminal) as a single mismatch, whether or not the gap was extended: 1.0- ((mismatches + gap openings)/(longest sequence length)).
\item (4) BLAST definition, equivalent to --iddef 1 for global pairwise alignments.
}}

\item{\code{rank_phylum}}{Character string containing Phylum rank taxonomy}

\item{\code{rank_class}}{Character string containing Class rank taxonomy}

\item{\code{rank_order}}{Character string containing Order rank taxonomy}

\item{\code{rank_family}}{Character string containing Family rank taxonomy}

\item{\code{rank_genus}}{Character string containing Genus rank taxonomy}

\item{\code{rank_species}}{Character string containing Species rank taxonomy}

\item{\code{phylum_threshold}}{Numeric string containing Phylum-level sequence similarity threshold for rank demarcation}

\item{\code{class_threshold}}{Numeric string containing Class-level sequence similarity threshold for rank demarcation}

\item{\code{order_threshold}}{Numeric string containing Order-level sequence similarity threshold for rank demarcation}

\item{\code{family_threshold}}{Numeric string containing Family-level sequence similarity threshold for rank demarcation}

\item{\code{genus_threshold}}{Numeric string containing Genus-level sequence similarity threshold for rank demarcation}

\item{\code{species_threshold}}{Numeric string containing Species-level sequence similarity threshold for rank demarcation}
}}

\seealso{
\code{\link{isoTAX}}
}
