% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{df_to_isoLIB}
\alias{df_to_isoLIB}
\title{Convert isoLIB .CSV output to isoLIB class object}
\usage{
df_to_isoLIB(df)
}
\arguments{
\item{df}{Dataframe in same format as .CSV output file from \code{\link{isoLIB}} step.}
}
\value{
Returns an S4 \code{\link{class-isoLIB}} object that can be used to generate interactive HTML output tables.
}
\description{
Helper function to convert isoLIB .CSV output to a \code{\link{class-isoLIB}} class object.
}
