% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{df_to_isoTAX}
\alias{df_to_isoTAX}
\title{Convert isoTAX .CSV output to isoTAX class object}
\usage{
df_to_isoTAX(df)
}
\arguments{
\item{df}{Dataframe in same format as .CSV output file from \code{\link{isoTAX}} step.}
}
\value{
Returns an S4 \code{\link{class-isoTAX}} object that can be used to generate interactive HTML output tables.
}
\description{
Helper function to convert isoTAX .CSV output to a \code{\link{class-isoTAX}} class object.
}
