% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{export_html}
\alias{export_html}
\alias{export_html-isoQC}
\alias{export_html-isoTAX}
\alias{export_html-isoLIB}
\title{Export HTML for isoQC > isoTAX > isoLIB class objects}
\usage{
\S4method{export_html}{isoQC}(obj)

\S4method{export_html}{isoTAX}(obj)

\S4method{export_html}{isoLIB}(obj)
}
\arguments{
\item{obj}{An S4 class object generated from one of \code{\link{isoQC}}, \code{\link{isoTAX}}, or \code{\link{isoLIB}} steps}
}
\value{
HTML output file saved to working directory.
}
\description{
S4 wrapper functions to export interactive HTML tables from \code{\link{isoQC}}, \code{\link{isoTAX}}, or \code{\link{isoLIB}} class objects. Saves to HTML to current working directory and automatically opens.
}
