% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db.R
\name{get_db}
\alias{get_db}
\title{Download taxonomic reference database}
\usage{
get_db(db = "16S_bac", force_update = FALSE, add_taxonomy = FALSE)
}
\arguments{
\item{db}{Database selection. One of "16S", "16S_arc", "18S", "ITS", or "cpn60"}

\item{force_update}{Forces new databases to be downloaded.}

\item{add_taxonomy}{Add full taxonomy to header to enable classification of sequences offline. This is primarily for cases of operation with no internet or when computing within an HPC cluster without administrator access. Results in a semicolon (;) delimited FASTA header as follows: >Accession_no;d__Domain;p__Phylum;c__Class;o__Order;f__Family;g__Genus;s__Species (e.g., >NR_042817.1;d__Bacteria;p__Verrucomicrobiota;c__Verrucomicrobiia;o__Verrucomicrobiales;f__Akkermansiaceae;g__Akkermansia;s__Akkermansia_muciniphila)}
}
\value{
Returns file path for database of interest
}
\description{
This function downloads taxonomic reference database and formats them for use.
}
\examples{
db.path <- get_db(db="16S", force_update=FALSE)
}
