% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Determine user operating system.}
\usage{
get_os()
}
\value{
Returns sysname as one of windows/osx-mac/linux
}
\description{
Determines the type of operating system being used.
}
\examples{
#Example 1 on a Windows-based operating system
os.index <- get_os()
print(os.index)

#Example 2 on a Mac operating system
os.index <- get_os()
print(os.index)

#Example 3 on a Linux operating system
os.index <- get_os()
print(os.index)
}
