% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sanger_date.R
\name{get_sanger_date}
\alias{get_sanger_date}
\title{get_sanger_date function}
\usage{
get_sanger_date(file = NULL)
}
\arguments{
\item{file}{The .ab1 file in from which to retrieve the date information. (Must be in S4 abif format)}
}
\value{
Returns date in "YYYY_MM_DD" format
}
\description{
Helper function to automatically retrieve run date from Sanger sequencing .ab1 files.
}
\examples{
#Path to the first listed .ab1 file in example directory
fpath <- file.path(system.file("extdata/abif_examples/rocket_salad", package = "isolateR"),
                   list.files(system.file("extdata/abif_examples/rocket_salad", package = "isolateR"))[1])
#Read in the ab1 file to S4 format
ab1.S4 <- sangerseqR::read.abif(fpath)

#Retrieve date
get_sanger_date(ab1.S4)
}
