% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vsearch.R
\name{get_vsearch}
\alias{get_vsearch}
\title{Download VSEARCH software reference database}
\usage{
get_vsearch(os = NULL)
}
\arguments{
\item{os}{Operating system, one of: "windows", "osx-mac", or "linux". If blank (os=NULL) then will try to automatically determine operating system.}
}
\value{
Returns path for VSEARCH executable
}
\description{
This function downloads the VSEARCH software used querying sequences against taxonomic databases of interest.
}
\examples{
#Example for automatically detecting operating system and downloading VSEARCH software
vsearch.path <- get_vsearch()
}
