% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoQC_function.R
\name{isoQC}
\alias{isoQC}
\title{Perform automated quality trimming of input .ab1 files}
\usage{
isoQC(
  input = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  export_fasta = TRUE,
  export_fasta_revcomp = FALSE,
  verbose = FALSE,
  exclude = NULL,
  min_phred_score = 20,
  min_length = 200,
  sliding_window_cutoff = NULL,
  sliding_window_size = 15,
  date = NULL,
  files_manual = NULL
)
}
\arguments{
\item{input}{Path to directory with .ab1 files.}

\item{export_html}{(Default=TRUE) Output the results as an HTML file}

\item{export_csv}{(Default=TRUE) Output the results as a CSV file.}

\item{export_fasta}{(Default=TRUE) Output the sequences in a FASTA file.}

\item{export_fasta_revcomp}{(Default=FALSE) Output the sequences in reverse complement form in a fasta file. This is useful in cases where sequencing was done using the reverse primer and thus the orientation of input sequences needs reversing.}

\item{verbose}{(Default=FALSE) Output progress while script is running.}

\item{exclude}{(Default=NULL) For testing purposes only. Excludes files of interest from input directory.}

\item{min_phred_score}{(Default=20) Do not accept trimmed sequences with a mean Phred score below this cutoff}

\item{min_length}{(Default=200) Do not accept trimmed sequences with sequence length below this number}

\item{sliding_window_cutoff}{(Default=NULL) Quality trimming parameter (M2) for wrapping SangerRead function in sangeranalyseR package. If NULL, implements auto cutoff for Phred score (recommended), otherwise set between 1-60.}

\item{sliding_window_size}{(Default=15) Quality trimming parameter (M2) for wrapping SangerRead function in sangeranalyseR package. Recommended range between 5-30.}

\item{date}{Set date "YYYY_MM_DD" format. If NULL, attempts to parse date from .ab1 file}

\item{files_manual}{(Default=NULL) For testing purposes only. Specify a list of files to run  as filenames without extensions, rather than the whole directory format. Primarily used for testing, use at your own risk.}
}
\value{
Returns quality trimmed Sanger sequences in FASTA format.
}
\description{
This function loads in ABIF files (.ab1 extension) and performs automatic quality trimming in batch mode.
}
\examples{
#Set path to directory containing example .ab1 files
fpath1 <- system.file("extdata/abif_examples/rocket_salad", package = "isolateR")

#Step 1: Run isoQC function with default settings
isoQC.S4 <- isoQC(input=fpath1)

#Show summary statistics
isoQC.S4
}
\seealso{
\code{\link{isoTAX}}, \code{\link{isoLIB}}
}
