% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoTAX_function.R
\name{isoTAX}
\alias{isoTAX}
\title{Classify taxonomy of sequences after quality trimming steps.}
\usage{
isoTAX(
  input = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  quick_search = TRUE,
  db = "16S",
  db_path = NULL,
  vsearch_path = NULL,
  iddef = 2,
  phylum_threshold = 75,
  class_threshold = 78.5,
  order_threshold = 82,
  family_threshold = 86.5,
  genus_threshold = 96.5,
  species_threshold = 98.7
)
}
\arguments{
\item{input}{Path of CSV output file from isoQC step.}

\item{export_html}{(Default=TRUE) Output the results as an HTML file}

\item{export_csv}{(Default=TRUE) Output the results as a CSV file.}

\item{quick_search}{(Default=FALSE) Whether or not to perform a comprehensive database search (i.e. optimal global alignment).
If TRUE, performs quick search equivalent to setting VSEARCH parameters "--maxaccepts 100 --maxrejects 100".
If FALSE, performs comprehensive search equivalent to setting VSEARCH parameters "--maxaccepts 0 --maxrejects 0"}

\item{db}{(Default="16S") Select database option(s) including "16S" (for searching against the NCBI Refseq targeted loci 16S rRNA database),
"ITS" (for searching against the NCBI Refseq targeted loci ITS  database. For combined databases in cases where input sequences are derived from
bacteria and fungi, select "16S|ITS". Setting to anything other than db=NULL or db="custom" causes 'db.path' parameter to be ignored.}

\item{db_path}{Path of FASTA-formatted database sequence file. Ignored if 'db' parameter is set to anything other than NULL or "custom".}

\item{vsearch_path}{Path of VSEARCH software if manually downloaded in a custom directory. If NULL (Default), will attempt automatic download.}

\item{iddef}{Set pairwise identity definition as per VSEARCH definitions (Default=2, and is recommended for highest taxonomic accuracy)
(0) CD-HIT definition: (matching columns) / (shortest sequence length).
(1) Edit distance: (matching columns) / (alignment length).
(2) Edit distance excluding terminal gaps (default definition).
(3) Marine Biological Lab definition counting each gap opening (internal or terminal) as a single mismatch, whether or not the gap was extended: 1.0- ((mismatches + gap openings)/(longest sequence length)).
(4) BLAST definition, equivalent to --iddef 1 for global pairwise alignments.}

\item{phylum_threshold}{Percent sequence similarity threshold for phylum rank demarcation}

\item{class_threshold}{Percent sequence similarity threshold for class rank demarcation}

\item{order_threshold}{Percent sequence similarity threshold for order rank demarcation}

\item{family_threshold}{Percent sequence similarity threshold for family rank demarcation}

\item{genus_threshold}{Percent sequence similarity threshold for genus rank demarcation}

\item{species_threshold}{Percent sequence similarity threshold for species rank demarcation}
}
\value{
Returns taxonomic classification table of class isoTAX. Default taxonomic cutoffs for phylum (75.0), class (78.5), order (82.0), family (86.5), genus (96.5), and species (98.7) demarcation are based on Yarza et al. 2014, Nature Reviews Microbiology (DOI:10.1038/nrmicro3330)
}
\description{
This function performs taxonomic classification steps by searching query Sanger sequences against specified database of interest. Takes CSV input files, extracts FASTA-formatted query sequences and performs global alignment against specified database of interest via Needleman-Wunsch algorithm by wrapping the --usearch_global command implemented in VSEARCH. Default taxonomic rank cutoffs for 16S rRNA gene sequences are based on Yarza et al. 2014, Nat Rev Microbiol.
}
\examples{
#Set path to directory containing example .ab1 files
fpath1 <- system.file("extdata/abif_examples/rocket_salad", package = "isolateR")

#Step 1: Run isoQC function with default settings
isoQC.S4 <- isoQC(input=fpath1)

#Step 2: Run isoTAX function with default settings
fpath2 <- file.path(fpath1, "isolateR_output/01_isoQC_trimmed_sequences_PASS.csv")
isoTAX.S4 <- isoTAX(input=fpath2)
#Show summary statistics
isoTAX.S4
}
\seealso{
\code{\link{isoQC}}, \code{\link{isoLIB}}, \code{\link{search_db}}
}
