% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fasta.R
\name{make_fasta}
\alias{make_fasta}
\title{Convert CSV file containing sequences to FASTA format}
\usage{
make_fasta(
  csv_file = NULL,
  col_names = "ID",
  col_seqs = "Sequence",
  output = "output.fasta"
)
}
\arguments{
\item{csv_file}{Filename (or path and filename if not in working directory) of the table from which you would like to generate a FASTA file.}

\item{col_names}{Column name with the unique names/identifiers. (Default="ID")}

\item{col_seqs}{Column name with the sequences. (Default="Sequence")}

\item{output}{Desired filename for output FASTA file (Default = "output.fasta")}
}
\value{
Returns sequences in FASTA format.
}
\description{
This function extracts sequences from a table in CSV format and converts them to FASTA format. Requires two columns, one with sequences and one with sequence names.
}
\examples{
#Set path to directory containing example .ab1 files
fpath1 <- system.file("extdata/abif_examples/rocket_salad", package = "isolateR")

#Run isoQC function with default settings to generate CSV file
isoQC.S4 <- isoQC(input=fpath1)

#Set path of CSV output file from isoQC step
csv.path <- file.path(fpath1, "isolateR_output/01_isoQC_trimmed_sequences_PASS.csv")

#Run make_fasta function
make_fasta(csv_file= csv.path, col_names="filename", col_seqs="seqs_trim", output="output.fasta")
}
