% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{make_tree}
\alias{make_tree}
\title{Generate a phylogenetic tree from an isoLIB output file}
\usage{
make_tree(input = NULL)
}
\arguments{
\item{input}{Full path to isoLIB strain library output file in .CSV format.}
}
\value{
Returns a \link[ggtree]{ggtree} class object
}
\description{
This script will help the user make a simple phylogenetic tree from a strain library.
It will allow the user to colour the tree by taxonomic rank only.
See \link[ggtree]{ggtree} documentation for more information on customization options available.
}
\examples{
#Set path to directory containing example .ab1 files
fpath1 <- system.file("extdata/abif_examples/rocket_salad", package = "isolateR")

#Step 1: Run isoQC function with default settings
isoQC.S4 <- isoQC(input=fpath1)

#Step 2: Run isoTAX function with default settings
fpath2 <- file.path(fpath1, "isolateR_output/01_isoQC_trimmed_sequences_PASS.csv")
isoTAX.S4 <- isoTAX(input=fpath2)

#Step 3: Run isoLIB function with default settings
fpath3 <- file.path(fpath1, "isolateR_output/02_isoTAX_results.csv")
isoLIB.S4 <- isoLIB(input=fpath3)
#Step 4: Make a tree from isoLIB output CSV file
fpath4 <- file.path(fpath1, "isolateR_output/03_isoLIB_results.csv")
make_tree(input= fpath4)
}
\seealso{
\code{\link{isoLIB}}
}
