% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{method-isoLIB}
\alias{method-isoLIB}
\title{setMethod functions for isoLIB}
\usage{
\S4method{isoLIB}{missing}(
  input = NULL,
  old_lib_csv = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  include_warnings = FALSE,
  strain_group_cutoff = 0.995,
  phylum_cutoff = 75,
  class_cutoff = 78.5,
  order_cutoff = 82,
  family_cutoff = 86.5,
  genus_cutoff = 96.5,
  species_cutoff = 98.7
)
}
\description{
Initiation of isoLIB functions.
}
