% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{method-isoQC}
\alias{method-isoQC}
\title{setMethod functions for isoQC}
\usage{
\S4method{isoQC}{missing}(
  input = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  export_fasta = TRUE,
  export_fasta_revcomp = FALSE,
  verbose = FALSE,
  exclude = NULL,
  min_phred_score = 20,
  min_length = 200,
  sliding_window_cutoff = NULL,
  sliding_window_size = 15,
  date = NULL,
  files_manual = NULL
)
}
\description{
Initiation of isoQC functions.
}
