% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{method-isoTAX}
\alias{method-isoTAX}
\title{setMethod functions for isoTAX}
\usage{
\S4method{isoTAX}{missing}(
  input = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  quick_search = TRUE,
  db = "16S_bac",
  db_path = NULL,
  vsearch_path = NULL,
  iddef = 2,
  phylum_threshold = 75,
  class_threshold = 78.5,
  order_threshold = 82,
  family_threshold = 86.5,
  genus_threshold = 94.5,
  species_threshold = 98.7
)
}
\description{
Initiation of isoTAX functions.
}
