% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{S4_to_dataframe}
\alias{S4_to_dataframe}
\title{Converts S4 objects (isoQC, isoTAX, or isoLIB) to dataframe}
\usage{
S4_to_dataframe(obj)
}
\arguments{
\item{obj}{S4 object generated from \code{\link{isoQC}}, \code{\link{isoTAX}}, or \code{\link{isoLIB}} steps}
}
\value{
Returns a dataframe containing sequence information in columns.
}
\description{
Helper function to convert S4 class objects (\code{\link{isoQC}}, \code{\link{isoTAX}}, or \code{\link{isoLIB}}) to dataframe
}
