% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanger_assembly.R
\name{sanger_assembly}
\alias{sanger_assembly}
\title{Overlap multiple paired Sanger sequences in batch.}
\usage{
sanger_assembly(input = NULL, suffix = "_F.ab1|_R.ab1")
}
\arguments{
\item{input}{Path of CSV output file from isoQC step.}

\item{suffix}{Regex-friendly suffix for denoting filename groupings. Default="_F.ab1|_R.ab1" for the common scenario of Sanger sequencing a marker gene in forward and reverse.
Direction of sequences including reverse complements will be automatically detected.}
}
\value{
Returns merged pairs of Sanger sequences in FASTA format.
}
\description{
This function loads in the CSV results table from isoQC and merges related sequences based on user input. Original file
names before isoQC step need to have a common prefix and differentiating suffixes. (e.g. SAMPLE_01_F.ab1, SAMPLE_01_R.ab1). After
aligning paired sequences, the consensus sequence is extracted and priority is given to the read with higher quality. Phred quality
scores are reassigned in the final output table in a basic way by taking the mean of both input sequences.

Note: This function is designed to be used after the isoQC step and before the isoTAX step.
}
\examples{
#Load package
library(isolateR)

#Step 1: Set path to directory containing paired .ab1 files
fpath <- system.file("extdata/abif_examples/drosophila_paired", package="isolateR")

#Step 2: Run isoQC function to trim poor quality regions (Phred score <20) before assembly
isoQC.S4 <- isoQC(input=fpath, sliding_window_cutoff = 20)

#Step 3: Assemble paired sequences
sanger_assembly(input = file.path(fpath,"isolateR_output", "01_isoQC_trimmed_sequences_PASS.csv"),
                suffix = "_F.ab1|_R.ab1")
                
#Detected 3 unique group(s) with suffix provided.
#Group    Individual filenames 
#-----    -------------------- 
# 1       DRO-1-isolate_F.ab1 | DRO-1-isolate_R.ab1 
# 2       DRO-2-isolate_F.ab1 | DRO-2-isolate_R.ab1 
# 3       DRO-3-isolate_F.ab1 | DRO-3-isolate_R.ab1

}
\seealso{
\code{\link{isoQC}}, \code{\link{isoTAX}}
}
