% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_tax_check.R
\name{valid_tax_check}
\alias{valid_tax_check}
\title{Validate species name via API client of LPSN}
\usage{
valid_tax_check(input = NULL, col_species = "rank_species", export_csv = TRUE)
}
\arguments{
\item{input}{CSV file path. Expects full path if CSV file is not in the current working directory.}

\item{col_species}{Specify the column containing the binomial species names (e.g. "Akkermansia muciniphila")}

\item{export_csv}{Toggle (TRUE/FALSE). Set TRUE to automatically write .CSV file of results to current directory. (Default=TRUE)}
}
\value{
Returns a CSV saved in working directory
}
\description{
This function will determine if each species in a CSV file is validly published or not.
Result file will be a CSV with the results appended to the input data.
This function requires the user to have an LPSN API account setup. For more details and to register, see here: https://api.lpsn.dsmz.de/)
}
