% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db.R
\name{get_db}
\alias{get_db}
\title{Download taxonomic reference database}
\usage{
get_db(db = "16S", force_update = FALSE)
}
\arguments{
\item{db}{Database selection. One of "16S", "16S_arc", "18S", "ITS", or "cpn60"}

\item{force_update}{Forces new datbases to be downloaded.}
}
\value{
Returns file path for database of interest
}
\description{
This function donwloads taxonomic reference database and formats them for use.
}
\examples{
db.path <- get_db(db="16S", force_update=FALSE)
}
