% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{method-isoLIB}
\alias{method-isoLIB}
\title{setMethod functions for isoLIB}
\usage{
\S4method{isoLIB}{missing}(
  input = NULL,
  old_lib_csv = NULL,
  method = "dark_mode",
  group_cutoff = 0.995,
  keep_old_reps = TRUE,
  export_html = TRUE,
  export_csv = TRUE,
  include_warnings = TRUE,
  vsearch_path = NULL,
  phylum_threshold = 75,
  class_threshold = 78.5,
  order_threshold = 82,
  family_threshold = 86.5,
  genus_threshold = 94.5,
  species_threshold = 98.7
)
}
\description{
Initiation of isoLIB functions.
}
