% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class_methods.R
\name{method-isoTAX}
\alias{method-isoTAX}
\title{setMethod functions for isoTAX}
\usage{
\S4method{isoTAX}{missing}(
  input = NULL,
  export_html = TRUE,
  export_csv = TRUE,
  quick_search = TRUE,
  database = "16S",
  iddef = 2,
  phylum_cutoff = 75,
  class_cutoff = 78.5,
  order_cutoff = 82,
  family_cutoff = 86.5,
  genus_cutoff = 96.5,
  species_cutoff = 98.7
)
}
\description{
Initiation of isoTAX functions.
}
